'give_drone' -> `spawn_minion`
'remove_all_drones` -> `remove_all_minions`
'give_drone' -> Replaced with `spawn_minion`

Maybe more agreeable overlap?
And built upon already better code instead of more so duplicate.

I just like being able to get the minion I want with 1 argument not 4 but that might not be too important.
But other proposed commands are of similiar stature.

Replaces elite with droneTier and axes the TeamIndex argument.
If any other minion specific stuff might get added it could also just get added to this command yknow.
